Create expand.dtfm()
Initial commit. 

This function takes a dataset from a data.frame() in which each row represents a certain covariate pattern with the number of observations meeting that pattern contained in a frequency column and expands the dataset creating a flat file dataset where each individual observation is contained in its own line.